+++ src/main/java/com/sparta/orderservice/payment/presentation/dto/response/ResPaymentDtoV1.java
@@ -6,9 +6,7 @@ import com.sparta.orderservice.payment.domain.entity.PaymentStatusEnum;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
@@ -21,7 +19,7 @@ public class ResPaymentDtoV1 {
 
     private UUID orderId;
 
-    private BigDecimal amount;
+    private Integer amount;
 
     private String payType; // 결제 방식 (카드, 계좌이체 등)
 
@@ -36,7 +34,7 @@ public class ResPaymentDtoV1 {
     private LocalDateTime canceledAt;
 
     @QueryProjection
-    public ResPaymentDtoV1(UUID paymentId, UUID orderId, BigDecimal amount, PaymentMethodEnum payType, String userName, PaymentStatusEnum status, LocalDateTime paidAt, LocalDateTime canceledAt, LocalDateTime updatedAt) {
+    public ResPaymentDtoV1(UUID paymentId, UUID orderId, Integer amount, PaymentMethodEnum payType, String userName, PaymentStatusEnum status, LocalDateTime paidAt, LocalDateTime canceledAt, LocalDateTime updatedAt) {
         this.paymentId = paymentId;
         this.orderId = orderId;
         this.amount = amount;
